+++ spine/model/layer/cnn/fpn.py
@@ -158,7 +158,7 @@ class FPN(torch.nn.Module):
 
     def forward(self, input):
         coords = input[:, 0:self.D+1].int()
-        features = input[:, self.D+1:].float()
+        features = input[:, self.D+1:]
 
         x = ME.SparseTensor(features, coords=coords)
         encoderOutput = self.encoder(x)
